Move remote util functions to manager and server
